feat: Move test_external_events to unit test dir
