+++ tests/unit/entrypoints/test_external_events.py
